services/product_service/requirements.txt
fastapi==0.109.1
pydantic==2.11.4
pydantic-settings==2.9.1
starlette==0.35.1

uvicorn[standard]==0.22.0
sqlalchemy==1.4.47

pyodbc
python-dotenv
